Don't use JavaScript modules anymore
They take time to load... and as they're loaded asyncronously, it was
making me lose some XHR requests, causing flickering.

Now, it works as it did pre-plugins. We generate one huge file and we
load it into a <script> tag. I'm writing the content right away because
adding it as a <script src> added a bit of delay.

This seems to be way better in terms of stability. The only thing I'm
going to miss is the possibility of doing dynamic module loading once FF
implemented it... but even in that scenario, the time to load it would
be bigger than the time it takes Slack to do their first request.